Change "green-background" id to "green-bold" class
In the previous step we skipped a step where we must bold the text of
the 4th element. To do this, we should rename the "green-background" id
to "green-bold". Changing it from an id to a class also would allow us
to possibly alter future elements with these same parameters.